// src/app/models/ResultEntry.ts
export class ResultEntry {
  id: number;
  question: number;
  answer: number;
  correct: boolean;
  date_added: string;
  comments: string;

  // constructor() {}

  // constructor(Id: number, Question: number, Date_Added: string) {
  //   this.id = Id;
  //   this.question = Question;
  //   this.date_added = Date_Added;

  // }
}